Move dependency bootstrap to app entry point
